requirements.txt
aiohttp~=3.9.2
pydantic~=2.5.3
pillow==10.2.0
aiogram==3.3.0
apscheduler==3.10.4
piccolo==1.3.0
pydantic-settings==2.1.0
piccolo[sqlite]
